requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Payment Gateway
razorpay==1.4.1

# Image Processing
Pillow==10.1.0

# CORS
fastapi-cors==0.0.6

# Email (optional)
fastapi-mail==1.4.1

# Testing
pytest==7.4.3
httpx==0.25.2





slowapi==0.1.9
redis==5.0.1
python-slugify==8.0.1

python-magic==0.4.27
bleach==6.1.0



python-slugify==8.0.1
slowapi==0.1.9
redis==5.0.1
celery==5.3.4
python-magic==0.4.27
bleach==6.1.0
jinja2==3.1.3
reportlab==4.0.8
